+++ tests/test_scrna/SeuratFilter/test.py
@@ -33,7 +33,9 @@ class SeuratFilter(SeuratFilter):
 
 
 def pipeline():
-    return get_pipeline(__file__).set_starts(PrepareSeurat)
+    return get_pipeline(__file__, plugins=["no:report"]).set_starts(
+        PrepareSeurat
+    )
 
 
 def testing(pipen):